Add support for reading data from logged buffers
When recording telemetry, buffers are appended to the end of the file.
Each buffer is the length described in the file header, and the number
of stored buffers are saved in the "disk subheader" (really just a part
of the header).

Also note that the API completely changed. A VarBuffer is now an object,
and the IRacingDataReader can query either the most recent VarBuffer or
the nth VarBuffer (in the case of logged buffers).